Working completion routine and support.
Potentially the case that most of the code I can find copied the example
completion routine (Toolbox Reference Volume 1, page 3-8), including
retrieving an extra byte off the end of the buffer.

Register 3 is 2 bytes but the returned size is 1 byte, and the 2nd byte
(if retrieved) varies on real hardware.